Switched to using RestartSource instead of actor supervisors.
