+++ proposals/streams/microprofile-streams-impl/src/main/java/org/eclipse/microprofile/streams/impl/StreamManager.java
@@ -1,16 +1,24 @@
 package org.eclipse.microprofile.streams.impl;
 
+import akka.Done;
+import akka.NotUsed;
 import akka.actor.*;
-import akka.kafka.ConsumerSettings;
-import akka.kafka.ProducerSettings;
-import akka.pattern.Backoff;
-import akka.pattern.BackoffOptions;
-import akka.stream.Materializer;
+import akka.kafka.*;
+import akka.kafka.javadsl.Consumer;
+import akka.kafka.javadsl.Producer;
+import akka.stream.*;
+import akka.stream.javadsl.Keep;
+import akka.stream.javadsl.RestartSource;
+import akka.stream.javadsl.Sink;
+import akka.stream.javadsl.Source;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.common.serialization.Deserializer;
 import org.apache.kafka.common.serialization.Serializer;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.apache.kafka.common.serialization.StringSerializer;
+import org.eclipse.microprofile.streams.Ack;
+import org.eclipse.microprofile.streams.Envelope;
 import scala.concurrent.duration.Duration;
 
 import javax.enterprise.context.ApplicationScoped;
@@ -20,7 +28,10 @@ import javax.json.bind.JsonbBuilder;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.util.Map;
+import java.util.concurrent.CompletionStage;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Function;
 
 /**
  * This starts subscribers and publishers.
@@ -78,10 +89,41 @@ public class StreamManager {
             // todo should also be configurable
             .withProperty(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
 
-    Props streamSupervisor = Props.create(ConsumerSupervisor.class,
-        () -> new ConsumerSupervisor<>(materializer, consumerSettings, descriptor, t));
+    // todo make exponential backoff parameters configurable
+    // todo name the right parts of the stream so that error reporting makes sense
+    Source<Done, NotUsed> stream = RestartSource.withBackoff(
+        Duration.create(3, TimeUnit.SECONDS),
+        Duration.create(30, TimeUnit.SECONDS),
+        0.2,
+        () -> {
+          AtomicReference<Function<Ack, CompletionStage<Void>>> ackFunction = new AtomicReference<>();
 
-    return startStream(streamSupervisor, descriptor);
+          akka.stream.javadsl.Flow<Envelope<Object>, Ack, Function<Ack, CompletionStage<Void>>> subscriber =
+              (akka.stream.javadsl.Flow) descriptor.incomingFlow(t);
+
+          return Consumer.committableSource(consumerSettings, Subscriptions.topics(descriptor.incomingTopic()))
+              .<Envelope<Object>>map(message -> new KafkaEnvelope<>(message, ackFunction.get()))
+              .viaMat(subscriber, (control, ack) -> {
+                ackFunction.set(ack);
+                return NotUsed.getInstance();
+              }).mapAsync(1, ack -> {
+            if (ack instanceof KafkaAck) {
+              return ((KafkaAck) ack).getMessage()
+                  .committableOffset().commitJavadsl().thenApply(d -> Done.getInstance());
+            } else {
+              throw new IllegalArgumentException("Don't know how to handle ack of type " + ack.getClass() +
+                  ". Kafka consumers must only emit acks returned by the passed in Envelope.");
+            }
+          });
+        }
+    );
+
+    KillSwitch killSwitch = stream
+        .viaMat(KillSwitches.single(), Keep.right())
+        .to(Sink.ignore())
+        .run(materializer);
+
+    return killSwitch::shutdown;
   }
 
   private <T> RunningStream startOutgoingStream(StreamDescriptor.SourceOutgoingDescriptor<? super T> descriptor, T t) {
@@ -107,20 +149,69 @@ public class StreamManager {
             // todo make configurable
             .withBootstrapServers("localhost:9092");
 
-    Props streamSupervisor = Props.create(ProducerSupervisor.class,
-        () -> new ProducerSupervisor<>(materializer, producerSettings, descriptor, t));
-
-    return startStream(streamSupervisor, descriptor);
-  }
-
-  private RunningStream startStream(Props streamSupervisor, StreamDescriptor<?> descriptor) {
     // todo make exponential backoff parameters configurable
-    BackoffOptions options = Backoff.onStop(streamSupervisor, "stream",
+    // todo name the right parts of the stream so that error reporting makes sense
+    Source<Done, NotUsed> stream = RestartSource.withBackoff(
         Duration.create(3, TimeUnit.SECONDS),
         Duration.create(30, TimeUnit.SECONDS),
-        0.2);
+        0.2,
+        () -> {
 
-    ActorRef actor = system.actorOf(options.props(), descriptor.id());
-    return () -> actor.tell(PoisonPill.getInstance(), ActorRef.noSender());
+          Source<Envelope<Object>, ?> publisher = (Source) descriptor.outgoingSource(t);
+
+          return publisher.map(envelope -> new ProducerMessage.Message<>(
+              new ProducerRecord<String, Object>(descriptor.outgoingTopic(), envelope.getPayload()),
+              envelope
+          ))
+              .via(Producer.flow(producerSettings))
+              .mapAsync(1, result -> result.message().passThrough().ack().thenApply(v -> Done.getInstance()));
+        });
+
+    KillSwitch killSwitch = stream
+        .viaMat(KillSwitches.single(), Keep.right())
+        .to(Sink.ignore())
+        .run(materializer);
+
+    return killSwitch::shutdown;
   }
+
+  private static class KafkaEnvelope<T> implements Envelope<T> {
+    private final ConsumerMessage.CommittableMessage<?, T> message;
+    private final Function<Ack, CompletionStage<Void>> ackFunction;
+
+    public KafkaEnvelope(ConsumerMessage.CommittableMessage<?, T> message,
+        Function<Ack, CompletionStage<Void>> ackFunction) {
+      this.message = message;
+      this.ackFunction = ackFunction;
+    }
+
+    @Override
+    public T getPayload() {
+      return message.record().value();
+    }
+
+    @Override
+    public CompletionStage<Void> ack() {
+      return ackFunction.apply(getAck());
+    }
+
+    @Override
+    public Ack getAck() {
+      return new KafkaAck(message);
+    }
+  }
+
+  private static class KafkaAck implements Ack {
+
+    private final ConsumerMessage.CommittableMessage<?, ?> message;
+
+    public KafkaAck(ConsumerMessage.CommittableMessage<?, ?> message) {
+      this.message = message;
+    }
+
+    public ConsumerMessage.CommittableMessage<?, ?> getMessage() {
+      return message;
+    }
+  }
+
 }
